Fixes import of modes (moves to util.js)
