+++ src/util.js
@@ -1,3 +1,8 @@
+export const PointFormMode = {
+  EDIT: 'EDIT',
+  ADD: 'ADD',
+};
+
 // Функция из интернета по генерации случайного числа из диапазона
 // Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
 export const getRandomInteger = (a = 0, b = 1) => {
